applying ns-3 coding style

+++ src/tor/model/tor-bktap.h
@@ -14,7 +14,6 @@
 #define VEGASBETA 6
 #define UDP_CELL_HEADER_SIZE (4 + 4 + 2 + 6 + 2 + 1)
 
-
 #define QUEUESTATE_STABLE 0
 #define QUEUESTATE_TIMEOUT 1
 #define QUEUESTATE_RECOVERY 2
@@ -358,7 +357,7 @@ public:
   map< uint32_t, Ptr<Packet> > cellMap;
 
   bool wasRetransmit;
-  uint32_t queueState;
+  uint8_t queueState;
 
   queue<uint32_t> ackq;
   queue<uint32_t> fwdq;
@@ -379,7 +378,7 @@ public:
     begRttSeq = 1;
     ssthresh = pow (2,10);
     dupackcnt = 0;
- 
+
     queueState = QUEUESTATE_STABLE;
   }
